tsdbv2: experimantal storage.Storage implementation
This commit adds a disk based alternantive implementation of of `tsdb`.It is based on
roaring bitmap  index found on [Featurebase](https://github.com/FeatureBaseDB/featurebase) formerly `pilosa`, on top of [Pebble](https://github.com/cockroachdb/pebble) key value store.

Key properties

- Multi tenant
- Disk based
- Sharded
- Order less ingestion ( samples and  exemplars can be appended in any order)
- Designed to handle open telemetry workloads. This includes removing all restrictions like label  size for exemplars etc. Fully open telemetry samples are ingested and queried without restriction.
- Efficient and fast high cardinality series handling. Care has been taken to ensure,
from storage perspective live or stale series have no difference, each series is lowevered
into a low cardinality bit, which is used by the index throughout the lifetime of the series.

This is initial implementation, focus is to be correct and have feature parity with `tsdb`.
There is no way to know if this is better or worse than `tsdb` yet, I havent ran benchmarks.
The idea is theoretically sound, but eventually we will have to test to see the results.

The end goal is to be a drop in replacement for `tsdb`.

Currently it is just a draft, I will remove the draft status when I am able  to replace
tsdb in `util/teststorage` package.

**Maintainers**: Please ignore this, I am publishing it so that I won't lose focus.
This is 2 weeks work  from idea to implementation. There is a long road ahead for it
to be ready for review.

Signed-off-by: Geofrey Ernest <geofreyernest@live.com>