chore: Sign project binary on MacOS
In order to be able to run the generated binary, it needs to be
"cosigned". Also, please note that we use the "codesign" binary that's
shipped in MacOS directly (instead of "promu codesign"), as that depends
on an image that doesn't support linux/arm64 (w.r.t. darwin's rosetta
emulation) yet:

```
WARNING: image platform (linux/amd64) does not match the expected platform (linux/arm64)
```

Signed-off-by: Pranshu Srivastava <rexagod@gmail.com>