chore(build): use bash script to calculate the extra docker tags needed
Currently when we push a docker image, we might inadvertently overwrite
the "latest" tag with a version that is breaking. For example releasing
2.53.4 meant that "latest" regressed from 3.2.1 to a non compatible
version.

The new script parses the version and checks for semver and returns tags
such as:
"latest" if the release is not a pre-release and its version is higher
  than all others.
"vX" if the release is not a pre-release and it is the latest in its
  major version.
"vX.Y" if the release is not a pre-release and it is the latest in its
  major+minor version.

Signed-off-by: György Krajcsovits <gyorgy.krajcsovits@grafana.com>