doc: Move library release tagging to a separate section
In RELEASE.md, the instructions to tag the library release used to be
inserted between the normal release tagging and an alternative way of
doing the release tagging. This was very confusing, os the alternative
way did _not_ include the library tagging.

I reckoned it is much less confusing to have a separate section for
the library tagging.

Also, replace the incorrect "optionally" by "alternatively".

Signed-off-by: beorn7 <beorn@grafana.com>