promqltest: make tests skippable by line
This PR passes TBRun down into the test struct and uses it to run the tests. This makes it possible
to skip them on a line/expression basis instead of filename basis.
This is useful for the Thanos Engine to skip select tests instead of needing to skip the whole file.

Signed-off-by: Michael Hoffmann <mhoffmann@cloudflare.com>