docs: Fix chunk format documentation for `varint` encoding
While preparing previous commits, we identified an inconsistency in the
chunk format documentation. The `varint` encoding can require up to 10
bytes for a 64-bit integer, such as when timestamps are encoded.
However, the chunk length field is a 32-bit integer, which requires at
most 5 bytes in `varint` encoding.

This is reflected in the code, where a maximum of 5 bytes are read when
parsing the chunk length.

    https://github.com/prometheus/prometheus/blob/50ba25f2735c1e33eeaa507207dc98bd596d58ac/tsdb/chunks/chunks.go#L709-L711

Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>
Signed-off-by: Victor Herrero Otal <victor.herrero.otal@sap.com>