POC: new native histogram extrapolation idea
Compensate for under/over estimating the left hand side (start side)
of buckets.
If the bucket level interpolation is inside the duration to
start, then extrapolate only to this extra duration on the left.
Otherwise if the duration to overall zero (from the overall count) is
inside the bucket's zero time, than the bucket's estimate might be too
much, calculate the increase from zero through the start time to get the
increase from the start to the first sample.

Signed-off-by: György Krajcsovits <gyorgy.krajcsovits@grafana.com>