fix(tsdb): appender does not honor append order
Fixes: #15177

Instead of 3 separate lists, keep a single list of samples with series
reference. To support different data types I did a union type where
floats are handled separately from complex types, but complex types are
kept in one place.

Not finished, in particular the WAL format is kept and there's a
conversion from the union type to separate record types for the WAL/WBL.

Signed-off-by: György Krajcsovits <gyorgy.krajcsovits@grafana.com>