promql: limit histogram extrapolation to assumed zero point
This is approach (3) in
https://github.com/prometheus/prometheus/issues/15976#issuecomment-3032095158

If we have limited the extrapolation of the count in a native
histogram, we manipulate the rest of the histogram around the
assumption that the zero point of the count is indeed the zero point
of the whole histogram. So we change the sum and all buckets including
the zero bucket in a way that it also reaches zero at the same point.
This keeps the histogram consistent by redistributing observations
between the buckets.

Cons I can think about right now:

- Our neat trick of ignoring incompatible bucket layouts of the 1st
  sample in case of a counter reset between the 1st and 2nd sample
  doesn't work anymore. (But it's probably not a big deal to drop this
  rarely relevant tweak.)

- If the extrapolation of count does _not_ go below zero, we do not
  change any of the other fields. However, they might still be
  extrapolated belew zero individually. This is maybe less of a deal
  with sum, as it could legitimately be below zero, but it will lead
  to exaggeration of some buckets in the result.

- The former point implies another weird behavior: If we change from a
  scenario where the count is very close to be extrapolated below zero
  to a scenario where the extrapolation of the count is limited by a
  small amount, the count will indeed only change by an infinitesimal
  amount. However, the individual buckets or the sum might change by a
  whole lot because suddenly we "switch on" the adjustment that was
  missing before.

Signed-off-by: beorn7 <beorn@grafana.com>